Crude but functional bug fix for rake.py
While using your implementation of Rake, I stumbled across a punctuation bug, where it didn't "ignore" the punctuation that I had set it to ignore. I found out that it was because the punctuations were adjacent to each other (ex. '],') so it did not return "False" when comparing with self.to_ignore ('],' does not equal ',' or ']'). I quickly wrote a new function to "replace" the lambda which did fix the problem. I am sure that you can come up with a better solution, but I just wanted to bring your attention to this bug.